Update src/drivers/schism/fabm_driver.h with MASK definitions (#64)
* Update src/drivers/schism/fabm_driver.h with MASK definitions

Recent versions of SCHISM need the MASK_TYPE and MASK_VALUE to be defined in the driver.  This Commit adds the relevant definitions.  We also update copyright and license (placed in public domain).

* Updated drivers/schism/fabm_driver with _FABM_HORIZONTAL_MASK_

SCHISM does not need a 3D mask, so for performance reasons `_FABM_HORIZONTAL_MASK_` should be defined

* Retracted addition of FABM_MASK_HORIZONTAL

Breaks existing code, we might reintroduce as soon as the 
1) discussion about performance is resolved
2) the SCHISM code respects this #define